+++ crimson_test/lib/twitter/util.dart
@@ -1,3 +1,4 @@
+import 'package:crimson/crimson.dart';
 import 'package:intl/intl.dart';
 
 DateTime? convertTwitterDateTime(String? twitterDateString) {
@@ -23,4 +24,18 @@ String formatTwitterDateString(String twitterDateString) {
     ..removeWhere((part) => part.startsWith('+'));
 
   return sanitized.join(' ');
+}
+
+class TwitterDateConverter extends JsonConverter<DateTime> {
+  const TwitterDateConverter();
+
+  @override
+  DateTime fromJson(dynamic json) {
+    return convertTwitterDateTime(json as String?)!;
+  }
+
+  @override
+  String toJson(DateTime object) {
+    return object.toIso8601String();
+  }
 }