+++ lib/crimson.dart
@@ -1,5 +1,5 @@
 library crimson;
 
-export 'src/annotations.dart';
+export 'src/annotations.dart' hide JsonIgnore, JsonKebabCase, JsonSnakeCase;
 export 'src/crimson.dart';
-export 'src/default_converters.dart';
+export 'src/json_converter.dart';